API_MANAGE/requirements.txt
sqlalchemy>=2.0
mysql-connector-python
requests>=2.31.0
bcrypt>=4.0.0
python-decouple>=3.8
prettytable>=3.10.0
